+++ geopackage-sdk/src/androidTest/java/mil/nga/geopackage/test/tiles/features/FeatureTileUtils.java
@@ -22,11 +22,11 @@ import mil.nga.geopackage.geom.GeoPackageGeometryData;
 import mil.nga.geopackage.tiles.features.DefaultFeatureTiles;
 import mil.nga.geopackage.tiles.features.FeatureTilePointIcon;
 import mil.nga.geopackage.tiles.features.FeatureTiles;
+import mil.nga.proj.ProjectionConstants;
 import mil.nga.sf.GeometryType;
 import mil.nga.sf.LineString;
 import mil.nga.sf.Point;
 import mil.nga.sf.Polygon;
-import mil.nga.sf.proj.ProjectionConstants;
 
 /**
  * Feature Tile Utils