+++ CHANGELOG.md
@@ -4,9 +4,14 @@ Adheres to [Semantic Versioning](http://semver.org/).
 
 ---
 
-## 5.1.1 (TBD)
+## [6.0.0](https://github.com/ngageoint/geopackage-android/releases/tag/6.0.0) (07-07-2021)
 
-* TBD
+* geopackage-core version 6.0.0
+* ormlite-android version 5.6
+* SQLite Android Bindings version 3360000
+* tiff version 2.0.3
+* appcompat version 1.3.0
+* gradle plugin updated to 4.2.2
 
 ## [5.1.0](https://github.com/ngageoint/geopackage-android/releases/tag/5.1.0) (03-23-2021)